deprecate cts resource and move to sub directory
